pyproject.toml
[project]
name = "workout-db"
version = "0.1.0"
description = "A simple messenger bot sending and receiving workout information stored in a database"
authors = [
    {name = "Tristan Wiessalla",email = "adstw@posteo.de"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "asyncpg>=0.27.0",  # PostgreSQL async library
    "python-telegram-bot>=20.0",  # Telegram Bot API
    "matrix-nio>=0.19.0",  # Matrix messaging API
    "python-dotenv>=1.0.0",  # Environment variable handling 
    "jupyter (>=1.1.1,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "notebook (>=7.3.2,<8.0.0)",
    "pyzmq (>=26.2.1,<27.0.0)",
    "jupyter-client (>=8.6.3,<9.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.poetry.group.dev.dependencies]
black = "^23.0"  # Code formatter
isort = "^5.12.0"  # Import sorter
mypy = "^1.2.0"  # Type checker
jupyter = "^1.1.1"
notebook = "^7.3.2"
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
pytest-mock = "^3.14.0"
coverage = "^7.6.12"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
